BD-244: Common remote (file) connector dialog does not test connection (e.g. in HDFS)
Task-Url: https://knime-com.atlassian.net/browse/BD-244

the test function now calls file.exists() to force the remote connection to be opened and validated. The problem with Saschas patch was that he called the exists method after the connection was closed via the monitor.closeAll() method. That is why I moved the monitor.closeAll() method to a finally block to have the connection closed after the checking and also if an exception is thrown. The closeAll() method is checking if a connection is opened and is also catching exceptions that might happen when closing the connection.