AP-13292: Use FileUtils to resolve paths
The ResolverUtil does not take care of UNC paths on windows (paths to
network drives which start with "\\"). Java has problems handling them
because there is no standard URI representation and they need to be
handled manually. This is done by the FileUtil class but not by the
ResolverUtil.

AP-13292 (Incompatibility of 'Binary Objects to Files' with network shares)