AP-16122: Do not try to force SSL session reuse when server does not support it
If the server does not support ssl session reuse between the control and data connections
then trying to reuse a the session from the control connection on the data connection,
will lead to the session being invalidated. This does not lead to an error directly,
but later attempts to reuse the invalidated session on the data connection
will result in the NPE described in the JIRA ticket.

AP-16122 (FTP Connectors causes NullPointerException in List Files/Folders node)